Extra Checkbox bug fixed
What was happening is that originally the document.querySelectorAll
was seeking out all of the inputs on the page. This included the
extra checkboxes on the page itself. Because each checkmark had a
name field with an empty string, it was being added as a key to the
object and then subsequently the checked portion was being assigned
to that key. Thus, we would have a "": true/false pair changing
as the loop would continue.

By specifically targetting the form via this.querySelectorAll we
now only scan through the original inputs on the form.